requirements.txt
flask
flask-compress
flask-cors
requests
python-Levenshtein
mysql-connector-python
Flask-SQLAlchemy
gunicorn
python-dotenv
PyJWT
flasgger 
pyyaml